Detect replaced Skill logs before resuming
